pyproject.toml
[project]
name = "oz-withpet"
version = "0.1.0"
description = ""
authors = [
    {name = "UNGBI78",email = "mybluewings78@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg (>=3.2.12,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

packages = [
    { include = "apps" }
]